+++ src/Parser.js
@@ -211,6 +211,7 @@ export default class Parser {
             if (!atom) {
                 break;
             } else if (atom.type === "internal") {
+                // Internal nodes do not appear in parse tree
                 continue;
             }
             body.push(atom);
@@ -286,7 +287,12 @@ export default class Parser {
         const symbol = symbolToken.text;
         this.consume();
         this.consumeSpaces(); // ignore spaces before sup/subscript argument
-        const group = this.parseGroup(name);
+
+        // Skip over allowed internal nodes such as \relax
+        let group: ?AnyParseNode;
+        do {
+            group = this.parseGroup(name);
+        } while (group?.type === "internal");
 
         if (!group) {
             throw new ParseError(
@@ -332,6 +338,12 @@ export default class Parser {
         // The body of an atom is an implicit group, so that things like
         // \left(x\right)^2 work correctly.
         const base = this.parseGroup("atom", breakOnTokenText);
+
+        // Internal nodes (e.g. \relax) cannot support super/subscripts.
+        // Instead we will pick up super/subscripts with blank base next round.
+        if (base?.type === "internal") {
+            return base;
+        }
 
         // In text mode, we don't have superscripts or subscripts
         if (this.mode === "text") {
@@ -668,7 +680,9 @@ export default class Parser {
         if (res == null) {
             return null;
         }
-        const match = (/^(#[a-f0-9]{3}|#?[a-f0-9]{6}|[a-z]+)$/i).exec(res.text);
+        const match = (
+            /^(#[a-f0-9]{3,4}|#[a-f0-9]{6}|#[a-f0-9]{8}|[a-f0-9]{6}|[a-z]+)$/i
+        ).exec(res.text);
         if (!match) {
             throw new ParseError("Invalid color: '" + res.text + "'", res);
         }