+++ src/domTree.js
@@ -17,6 +17,7 @@ import {path} from "./svgGeometry";
 import type Options from "./Options";
 import {DocumentFragment} from "./tree";
 import {makeEm} from "./units";
+import ParseError from "./ParseError";
 
 import type {VirtualNode} from "./tree";
 
@@ -84,6 +85,16 @@ const toNode = function(tagName: string): HTMLElement {
 };
 
 /**
+ * https://w3c.github.io/html-reference/syntax.html#syntax-attributes
+ *
+ * > Attribute Names must consist of one or more characters
+ * other than the space characters, U+0000 NULL,
+ * '"', "'", ">", "/", "=", the control characters,
+ * and any characters that are not defined by Unicode.
+ */
+const invalidAttributeNameRegex = /[\s"'>/=\x00-\x1f]/;
+
+/**
  * Convert into an HTML markup string
  */
 const toMarkup = function(tagName: string): string {
@@ -110,6 +121,9 @@ const toMarkup = function(tagName: string): string {
     // Add the attributes
     for (const attr in this.attributes) {
         if (this.attributes.hasOwnProperty(attr)) {
+            if (invalidAttributeNameRegex.test(attr)) {
+                throw new ParseError(`Invalid attribute name '${attr}'`);
+            }
             markup += ` ${attr}="${utils.escape(this.attributes[attr])}"`;
         }
     }
@@ -217,7 +231,7 @@ export class Span<ChildType: VirtualNode> implements HtmlDomNode {
     }
 
     hasClass(className: string): boolean {
-        return utils.contains(this.classes, className);
+        return this.classes.includes(className);
     }
 
     toNode(): HTMLElement {
@@ -258,7 +272,7 @@ export class Anchor implements HtmlDomNode {
     }
 
     hasClass(className: string): boolean {
-        return utils.contains(this.classes, className);
+        return this.classes.includes(className);
     }
 
     toNode(): HTMLElement {
@@ -294,7 +308,7 @@ export class Img implements VirtualNode {
     }
 
     hasClass(className: string): boolean {
-        return utils.contains(this.classes, className);
+        return this.classes.includes(className);
     }
 
     toNode(): Node {
@@ -396,7 +410,7 @@ export class SymbolNode implements HtmlDomNode {
     }
 
     hasClass(className: string): boolean {
-        return utils.contains(this.classes, className);
+        return this.classes.includes(className);
     }
 
     /**